+++ src/main/java/com/supermartijn642/landmines/Landmines.java
@@ -6,7 +6,7 @@ import com.supermartijn642.core.registry.GeneratorRegistrationHandler;
 import com.supermartijn642.core.registry.RegistrationHandler;
 import com.supermartijn642.core.registry.RegistryEntryAcceptor;
 import com.supermartijn642.landmines.generators.*;
-import net.minecraft.resources.ResourceLocation;
+import net.minecraft.resources.Identifier;
 import net.minecraft.sounds.SoundEvent;
 import net.minecraftforge.fml.common.Mod;
 
@@ -37,7 +37,7 @@ public class Landmines {
             handler.registerItemCallback(type::registerItem);
         }
         // Trigger sound
-        handler.registerSoundEvent("trigger_sound", SoundEvent.createVariableRangeEvent(ResourceLocation.fromNamespaceAndPath("landmines", "trigger_sound")));
+        handler.registerSoundEvent("trigger_sound", SoundEvent.createVariableRangeEvent(Identifier.fromNamespaceAndPath("landmines", "trigger_sound")));
     }
 
     private static void registerGenerators(){