+++ changelog.md
@@ -1,2 +1,2 @@
 ### Landmines 1.0.7
-- Initial release of Landmines for Minecraft 1.21.9 & 1.21.10
+- Initial release of Landmines for Minecraft 1.21.11